guest

+++ client/src/Components/Carrito/Carrito.js
@@ -15,7 +15,7 @@ const useStyles = makeStyles((theme) => ({
 }));
 const Carrito = () => {
     const user = useSelector(store => store.user.user)
-
+    const [product, setProduct] = useState([]);
     const classes = useStyles();
     const carrito = useSelector(store => store.carrito.carrito) //Accedo al estado del carrito
     const dispatch = useDispatch()
@@ -28,23 +28,68 @@ const Carrito = () => {
         //   await axios.post(`http://localhost:4000/users/${user.id}/carrito`)    
         // }
         // fetchData()
+        if (!localStorage.token && localStorage.carrito) {
+          let data = JSON.parse(localStorage.getItem("carrito"));
+          return setProduct(data.carrito);
+      }
+      
     },[carrito])
 
     const DestroyCart = async()=>{ //Vacía el carrito
+      if (!localStorage.token) {
+        if (localStorage.carrito) {
+            setProduct([])
+            return localStorage.clear();
+        }
+    }
      const {data} = await axios.delete(`http://localhost:4000/users/${user.id}/deletecart/${carrito.id}`)
      alert('Carrito eliminado correctamente')
     }
 
     const handleBuy = async() =>{ // Manejador de la compra: cambia el status de "creado" a "procesando"
-    const {data} = await axios.put(`http://localhost:4000/users/procesando/${carrito.id}`) 
-    alert('Compra exitosa')
+    if(!user.id){
+      alert('Debes logearte primero')
+    }
+
+    if(user.id){
+      const {data} = await axios.put(`http://localhost:4000/users/procesando/${carrito.id}`) 
+      alert('Compra exitosa')
+    }
+    
+   
     }
 
     const DeleteProduct = async(id)=>{ //  Elimina el producto
-     const {data} = await axios.delete(`http://localhost:4000/users/product/${id}/delete/${carrito.id}`)
+      
+      if (!localStorage.token) {
+        if (localStorage.carrito) {
+            console.log("esta entrando")
+            let data = JSON.parse(localStorage.getItem("carrito"));
+            console.log(data.carrito)
+            console.log(data.carrito.productId)
+            console.log(id)
+            let otradata = data.carrito.filter(p => p.productId !== id);
+            localStorage.setItem("carrito", JSON.stringify({ carrito: otradata }))
+            return setProduct(otradata);
+        }
+
     }
+    if (localStorage.carrito) {
+      console.log("esta entrando")
+      let dato = JSON.parse(localStorage.getItem("carrito"));
+      console.log(dato.carrito)
+      console.log(dato.carrito.productId)
+      console.log(id)
+      let otradata = dato.carrito.filter(p => p.productId !== id);
+      localStorage.setItem("carrito", JSON.stringify({ carrito: otradata }))
+      const {data} = await axios.delete(`http://localhost:4000/users/product/${id}/delete/${carrito.id}`)
+  }
+     
+    } 
 
     const Increment = async(id, stock)=>{ //Realiza el put, y aumenta la cantidad del mismo producto
+
+
        if(stock<=1){
          alert ('No hay stock suficiente')
        }else if (stock > 0){
@@ -65,7 +110,34 @@ const Carrito = () => {
 
     return( 
         <div>
-       { carrito.products && carrito.products[0] && <div><table class="table table-striped table-dark">
+            { product && product[0] && <div><table class="table table-striped table-dark">
+  <thead>
+    <tr>
+      <th scope="col">Product</th>
+      <th scope="col">Price</th>
+      <th scope="col">Stock</th>
+      <th scope="col">Quantity</th>
+      <th scope="col">Total</th>
+      <th scope="col"></th>
+    </tr>
+  </thead>
+  <tbody>
+      {product && product.map((e)=>(
+        <tr>
+         <td>{e.name}</td>
+         <td>{e.price}</td>
+         <td>{e.stock}</td>
+         <td >{e.quantity}</td>
+        
+      <td ><label for={e.name}>${e.price * e.quantity }</label></td>
+         <td scope="col"> <Button variant="contained" color="secondary" onClick={()=> DeleteProduct(e.productId)} >Borrar</Button></td>
+        </tr>
+      ))}
+  </tbody>
+</table>
+<Button variant="contained" style={{backgroundColor: "green", color: "white", marginLeft: "10px"}} onClick={()=> handleBuy()} >Comprar</Button> 
+<Button variant="contained" style={{backgroundColor: "red", color: "white", marginLeft: "30px"}} onClick={()=> DestroyCart()} >Borrar carrito</Button></div>}     
+       { user.id && carrito.products && carrito.products[0] && <div><table class="table table-striped table-dark">
   <thead>
     <tr>
       <th scope="col">Product</th>
@@ -78,7 +150,7 @@ const Carrito = () => {
     </tr>
   </thead>
   <tbody>
-      {carrito.products && carrito.products.map((e)=>(
+      { user.id && carrito.products && carrito.products.map((e)=>(
         <tr>
          <td>{e.name}</td>
          <td>{e.price}</td>
@@ -93,11 +165,16 @@ const Carrito = () => {
 </table>
 <Button variant="contained" style={{backgroundColor: "green", color: "white", marginLeft: "10px"}} onClick={()=> handleBuy()} >Comprar</Button> 
 <Button variant="contained" style={{backgroundColor: "red", color: "white", marginLeft: "30px"}} onClick={()=> DestroyCart()} >Borrar carrito</Button></div>}
-{!carrito && <div className='titnocarrito' style={{marginLeft: "340px"}}>
+{user.id && !carrito.products[0] && <div className='titnocarrito' style={{marginLeft: "340px"}}>
                     <div className='divcarritovacio'>
                     <h2>¡Tu carrito está vacío!</h2>
                     </div>
                 </div>}
+{ !user.id && !product[0]  && <div className='titnocarrito' style={{marginLeft: "340px"}}>
+                    <div className='divcarritovacio'>
+                    <h2>¡Tu carrito está vacío!</h2>
+                    </div>
+                </div>}                
         </div>
     )
 }
